# apps/api/openapi.yaml
openapi: 3.1.0
info:
  title: wstprtradio API
  version: 0.0.1
  description: |
    Fastify control-plane API for West Port Radio.

    This specification is intentionally concise and human-maintained during the
    early build-out stage. It covers the endpoints needed for local deployment,
    operator troubleshooting, and future GitHub-hosted docs.
servers:
  - url: http://localhost:3001
    description: Local API
  - url: https://api.wstprtradio.com
    description: Production API
...
components:
  securitySchemes:
    sessionCookie:
      type: apiKey
      in: cookie
      name: sessionId
  schemas:
...
    Destination:
      type: object
      properties:
        id:
          type: string
        kind:
          type: string
        name:
          type: string
        enabled:
          type: boolean
        url:
          type: string
        sortOrder:
          type: integer
        createdAt:
          type: string
        updatedAt:
          type: string
      required: [id, kind, name, enabled, url, sortOrder, createdAt, updatedAt]
    DestinationCreate:
      type: object
      properties:
        kind:
          type: string
          enum: [twitch, instagram, custom_rtmp, custom_srt, tiktok_experimental, discord_notify]
        name:
          type: string
        enabled:
          type: boolean
        url:
          type: string
        streamKey:
          type: string
        srtPassphrase:
          type: string
        sortOrder:
          type: integer
      required: [kind, name, url]
...
paths:
...
  /admin/destinations:
    get:
      tags: [Admin]
      summary: List fanout destinations
      security:
        - sessionCookie: []
      responses:
        '200':
          description: Destination list.
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/Destination'
    post:
      tags: [Admin]
      summary: Create a destination
      security:
        - sessionCookie: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/DestinationCreate'
      responses:
        '201':
          description: Destination created.
        '400':
          description: Invalid body.
  /admin/destinations/{id}:
    patch:
      tags: [Admin]
      summary: Update a destination
      security:
        - sessionCookie: []
      parameters:
        - in: path
          name: id
          required: true
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/DestinationCreate'
      responses:
        '200':
          description: Destination updated.
        '404':
          description: Destination not found.
    delete:
      tags: [Admin]
      summary: Delete a destination
      security:
        - sessionCookie: []
      parameters:
        - in: path
          name: id
          required: true
          schema:
            type: string
      responses:
        '200':
          description: Destination deleted.
        '404':
          description: Destination not found.
  /admin/destinations/{id}/test:
    post:
      tags: [Admin]
      summary: Test destination connectivity via Cloudflare output creation
      security:
        - sessionCookie: []
      parameters:
        - in: path
          name: id
          required: true
          schema:
            type: string
      responses:
        '200':
          description: Connectivity test result.
        '404':
          description: Destination not found.
...